requirements.txt
flask
flask_cors
accelerate
tenacity
google-generativeai
ratelimit